requirements.txt
streamlit
pandas
numpy
openpyxl
matplotlib
datetime
Pillow
pyautogui